Changed a mustache file, added campaign routes
This branch us created from the delete-asset-test branch. On that branch I modified the api.mustache file and that change had a bug so I changed this file again on this branch in order to revert it to the initial state. I added routes for working with Campaign objects (for creating/returning a Campaign object)