Add campaign routes (#3)
* Added the Id property for the asset object and wrote the DeleteAssetTest

* Removed unused variable

* Worked on adding campaign routes

* Refactored the DeleteAssetTest

* Replaced Asset class name with {{classname}} in the api.mustache file

* Changed a mustache file, added campaign routes

This branch us created from the delete-asset-test branch. On that branch I modified the api.mustache file and that change had a bug so I changed this file again on this branch in order to revert it to the initial state. I added routes for working with Campaign objects (for creating/returning a Campaign object)

* This commit reverts a change made on the api.mustache file in a previous commit

This commit reverts a change made on the api.mustache file in a previous commit because that change was not helpful.

* Deleted two .cs files and renamed a test function.

Deleted two .cs files that were the result of code generation over a typo in the OpenAPI spec. Changed the name of the DeleteAssetTest to DeleteAssetByIdTest.

* Implemented the required changes

Created factories for Campaign and Asset objects, removed the logic for eliminating white spaces in the CampaignApiTests.cs, created the DeleteNonExistingAssetTest

* Refactored three tests in AssetApiTests.cs

Refactored three tests from AssetApiTests.cs so that the posted object will be deleted if one of the asserts failed.